Improve geosop performance, refactor
